+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.3.0]
+### Added
+- Added FieldOfView.
+- Added Flip functions to Cone and Pyramid
+
+### Fixed
+- Added missing axis support for CapsuleCollider drawing.
+- Fixed incorrect calculations for Horizontal CapsuleCollider2D drawing.
+
 ## [2.2.0]
 ### Added
 - Added PolygonCollider2D support to Shape drawing.